requirements.txt
activate==0.2.1
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
confusable-homoglyphs==3.0.0
cryptography==3.1.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2
django-allauth==0.42.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-datatable==0.3.1
django-datatables-ajax==0.8
django-filter==2.3.0
django-otp==1.0.0
django-postgresql==0.0.3
django-registration==3.1
django-rest-knox==4.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
djangorestframework-datatables==0.5.2
djangorestframework-simplejwt==4.4.0
factory-boy==2.10.0
Faker==4.1.3
filelock==3.0.12
gunicorn==19.7.1
idna==2.10
import==0.0.1
include==0.2.2
ipaddress==1.0.19
isort==5.4.2
lazy-object-proxy==1.4.3
Library==0.0.0
Markdown==3.2.2
mccabe==0.6.1
numpy==1.19.1
oauthlib==3.1.0
path==15.0.0
Pillow==7.2.0
pipenv==2020.8.13
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.1
python-dateutil==2.7.2
python-http-client==3.0.0
python3-openid==3.2.0
pytz==2018.7
qrcode==6.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
twilio==6.45.3
urllib3==1.25.10
views==0.3
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
whitenoise==3.3.1
wrapt==1.12.1
